requirements.txt
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-community>=0.0.10
requests>=2.31.0
python-dotenv>=1.0.0
PyGithub>=1.58.0
click>=8.1.0
pydantic>=2.0.0
colorlog>=6.7.0 